refactor: move TableTags para common/frontend
